LP-922 extract getPartners and partnerDateFormat to util
